refactor(server): split scheduler.go into 4 domain files (STRUCT-11)
